feat(1to1): rkgame 1:1 忠实重构地基 —— 差分验证门禁 + 源树确权 + 重建台账
P1-A 差分验证工具链（已自检 golden-vs-golden = 100% T1，扰动可拦截）:
  - tools/funcdump.py   objdump -d -> 逐函数归一化指令序列（消除布局差异）
  - tools/funcdiff.py   函数级门禁 T1/T2/T3/FAIL，输出语义等价覆盖率
  - golden/factory.funcs.json  原厂金标准（835 函数 / 715,473 指令）

P1-B 源树还原 + 确权（41 个 STT_FILE 编译单元）:
  - 上游开源 581 函数/194,454 B (61.3%)，专有 223 函数/122,922 B (38.7%)
  - 更正：'zlib 1.2.5' 系误判（是 .rodata 数据表字节）；
    XUnzip 桶实为 XZip(TUnzip)+miniunz+内嵌 zlib 1.1.x（huft_build/inflate_blocks 特征）
  - ledger/functions.csv  223 个专有函数逐函数台账（含模块/尺寸/源文件）

P2-A 版本指纹匹配工具（vermatch.py / stb_range.py）:
  - stb_truetype 名称指纹在 v1.19~v1.26 全区间 81/81 覆盖
    => 名称不足以定版，须同工具链编译后按函数尺寸/指令序列落定

+++ 1to1/src/proprietary/misc/FUN_00014470_UnDrawSelectBar.c
@@ -20,7 +20,7 @@ void UnDrawSelectBar(int *param_1,int param_2,int param_3)
   int iVar7;
   int iVar8;
   
-  iVar2 = DAT_003af29c;
+  iVar2 = (int)DAT_003af29c;
   iVar7 = param_1[1];
   if (param_1[3] <= iVar7) {
     return;
